XRDDEV-2189 separated management requests into separate commons
