+++ src/centralserver/admin-service/core-api/build.gradle
@@ -8,6 +8,7 @@ dependencies {
     api project(':signer-protocol')
     api project(':common-verifier')
     api project(':common-rest-api') //TODO should be decoupled
+    api project(':common-management-request')
 
     api('org.springframework.data:spring-data-commons') // TODO used for pagination, create DTOs.
