+++ src/centralserver/admin-service/core-api/src/main/java/org/niis/xroad/cs/admin/api/domain/AuthenticationCertificateRegistrationRequest.java
@@ -31,7 +31,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.Accessors;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.domain.Origin;
 
 @Getter
