+++ src/centralserver/admin-service/core-api/src/main/java/org/niis/xroad/cs/admin/api/domain/ClientDeletionRequest.java
@@ -29,7 +29,7 @@ package org.niis.xroad.cs.admin.api.domain;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.domain.Origin;
 
 @Getter
