+++ src/centralserver/admin-service/core-api/src/main/java/org/niis/xroad/cs/admin/api/domain/OwnerChangeRequest.java
@@ -30,7 +30,7 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.domain.Origin;
 
 import java.beans.Transient;
