+++ src/centralserver/admin-service/core-api/src/main/java/org/niis/xroad/cs/admin/api/dto/ManagementRequestInfoDto.java
@@ -30,7 +30,7 @@ import ee.ria.xroad.common.identifier.SecurityServerId;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.niis.xroad.centralserver.restapi.domain.ManagementRequestStatus;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.domain.Origin;
 
 import java.time.Instant;
