+++ src/centralserver/admin-service/core/src/main/java/org/niis/xroad/centralserver/restapi/dto/AuthenticationCertificateDeletionRequestDto.java
@@ -30,8 +30,8 @@ import ee.ria.xroad.common.identifier.SecurityServerId;
 
 import lombok.Getter;
 import org.niis.xroad.centralserver.restapi.domain.ManagementRequestStatus;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.domain.Origin;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 
 @Getter
 public class AuthenticationCertificateDeletionRequestDto extends ManagementRequestDto {
