+++ src/centralserver/admin-service/core/src/main/java/org/niis/xroad/cs/admin/core/entity/ClientRegistrationRequestEntity.java
@@ -30,8 +30,8 @@ import ee.ria.xroad.common.identifier.SecurityServerId;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.domain.Origin;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 
 import javax.persistence.Access;
 import javax.persistence.AccessType;
