+++ src/centralserver/admin-service/core/src/main/java/org/niis/xroad/cs/admin/core/entity/RequestEntity.java
@@ -30,8 +30,8 @@ package org.niis.xroad.cs.admin.core.entity;
 import lombok.Getter;
 import lombok.Setter;
 import org.niis.xroad.centralserver.restapi.domain.ManagementRequestStatus;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.domain.Origin;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 
 import javax.persistence.Column;
 import javax.persistence.DiscriminatorColumn;
