+++ src/centralserver/admin-service/infra-api-rest/src/main/java/org/niis/xroad/centralserver/restapi/converter/db/ManagementRequestDtoConverter.java
@@ -40,8 +40,8 @@ import org.niis.xroad.centralserver.openapi.model.ManagementRequestsFilterDto;
 import org.niis.xroad.centralserver.restapi.converter.model.ManagementRequestDtoTypeConverter;
 import org.niis.xroad.centralserver.restapi.converter.model.ManagementRequestOriginDtoConverter;
 import org.niis.xroad.centralserver.restapi.converter.model.ManagementRequestStatusConverter;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.dto.converter.DtoConverter;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 import org.niis.xroad.cs.admin.api.domain.AuthenticationCertificateDeletionRequest;
 import org.niis.xroad.cs.admin.api.domain.AuthenticationCertificateRegistrationRequest;
 import org.niis.xroad.cs.admin.api.domain.ClientDeletionRequest;
