+++ src/centralserver/admin-service/infra-api-rest/src/main/java/org/niis/xroad/centralserver/restapi/converter/model/ManagementRequestDtoTypeConverter.java
@@ -33,7 +33,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.ValueMapping;
 import org.mapstruct.ValueMappings;
 import org.niis.xroad.centralserver.openapi.model.ManagementRequestTypeDto;
-import org.niis.xroad.centralserver.restapi.domain.ManagementRequestType;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 import org.niis.xroad.centralserver.restapi.dto.converter.DtoConverter;
 
 import javax.persistence.AttributeConverter;
