+++ src/common-management-request/src/main/java/org/niis/xroad/common/managementrequest/ManagementRequestBuilder.java
@@ -4,17 +4,17 @@
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
  * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,7 +23,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package ee.ria.xroad.common.request;
+package org.niis.xroad.common.managementrequest;
 
 import ee.ria.xroad.common.identifier.ClientId;
 import ee.ria.xroad.common.identifier.SecurityServerId;
@@ -32,8 +32,13 @@ import ee.ria.xroad.common.message.ProtocolVersion;
 import ee.ria.xroad.common.message.SoapBuilder;
 import ee.ria.xroad.common.message.SoapHeader;
 import ee.ria.xroad.common.message.SoapMessageImpl;
+import ee.ria.xroad.common.request.AuthCertDeletionRequestType;
+import ee.ria.xroad.common.request.AuthCertRegRequestType;
+import ee.ria.xroad.common.request.ClientRequestType;
+import ee.ria.xroad.common.request.ObjectFactory;
 
 import lombok.extern.slf4j.Slf4j;
+import org.niis.xroad.common.managementrequest.model.ManagementRequestType;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBElement;
@@ -42,11 +47,11 @@ import javax.xml.namespace.QName;
 
 import java.util.UUID;
 
-import static ee.ria.xroad.common.request.ManagementRequests.AUTH_CERT_DELETION;
-import static ee.ria.xroad.common.request.ManagementRequests.AUTH_CERT_REG;
-import static ee.ria.xroad.common.request.ManagementRequests.CLIENT_DELETION;
-import static ee.ria.xroad.common.request.ManagementRequests.CLIENT_REG;
-import static ee.ria.xroad.common.request.ManagementRequests.OWNER_CHANGE;
+import static org.niis.xroad.common.managementrequest.model.ManagementRequestType.AUTH_CERT_DELETION_REQUEST;
+import static org.niis.xroad.common.managementrequest.model.ManagementRequestType.AUTH_CERT_REGISTRATION_REQUEST;
+import static org.niis.xroad.common.managementrequest.model.ManagementRequestType.CLIENT_DELETION_REQUEST;
+import static org.niis.xroad.common.managementrequest.model.ManagementRequestType.CLIENT_REGISTRATION_REQUEST;
+import static org.niis.xroad.common.managementrequest.model.ManagementRequestType.OWNER_CHANGE_REQUEST;
 
 @Slf4j
 final class ManagementRequestBuilder {
@@ -72,7 +77,7 @@ final class ManagementRequestBuilder {
         request.setAddress(address);
         request.setAuthCert(authCert);
 
-        return buildMessage(element(AUTH_CERT_REG, AuthCertRegRequestType.class, request));
+        return buildMessage(element(AUTH_CERT_REGISTRATION_REQUEST, AuthCertRegRequestType.class, request));
     }
 
     SoapMessageImpl buildAuthCertDeletionRequest(SecurityServerId.Conf securityServer,
@@ -83,7 +88,7 @@ final class ManagementRequestBuilder {
         request.setServer(securityServer);
         request.setAuthCert(authCert);
 
-        return buildMessage(element(AUTH_CERT_DELETION, AuthCertDeletionRequestType.class, request));
+        return buildMessage(element(AUTH_CERT_DELETION_REQUEST, AuthCertDeletionRequestType.class, request));
     }
 
     SoapMessageImpl buildClientRegRequest(SecurityServerId.Conf securityServer, ClientId.Conf client) throws Exception {
@@ -93,7 +98,7 @@ final class ManagementRequestBuilder {
         request.setServer(securityServer);
         request.setClient(client);
 
-        return buildMessage(element(CLIENT_REG, ClientRequestType.class, request));
+        return buildMessage(element(CLIENT_REGISTRATION_REQUEST, ClientRequestType.class, request));
     }
 
     SoapMessageImpl buildClientDeletionRequest(SecurityServerId.Conf securityServer,
@@ -104,7 +109,7 @@ final class ManagementRequestBuilder {
         request.setServer(securityServer);
         request.setClient(client);
 
-        return buildMessage(element(CLIENT_DELETION, ClientRequestType.class, request));
+        return buildMessage(element(CLIENT_DELETION_REQUEST, ClientRequestType.class, request));
     }
 
     SoapMessageImpl buildOwnerChangeRequest(SecurityServerId.Conf securityServer,
@@ -115,7 +120,7 @@ final class ManagementRequestBuilder {
         request.setServer(securityServer);
         request.setClient(client);
 
-        return buildMessage(element(OWNER_CHANGE, ClientRequestType.class, request));
+        return buildMessage(element(OWNER_CHANGE_REQUEST, ClientRequestType.class, request));
     }
 
     // -- Private helper methods ----------------------------------------------
@@ -149,8 +154,8 @@ final class ManagementRequestBuilder {
         return JAXB_CTX.createMarshaller();
     }
 
-    private static <T> JAXBElement<T> element(String name, Class<T> clazz, T value) {
-        return new JAXBElement<>(new QName(SoapHeader.NS_XROAD, name), clazz, null, value);
+    private static <T> JAXBElement<T> element(ManagementRequestType requestType, Class<T> clazz, T value) {
+        return new JAXBElement<>(new QName(SoapHeader.NS_XROAD, requestType.getServiceCode()), clazz, null, value);
     }
 
     private static JAXBContext initJaxbContext() {
