+++ src/common-management-request/src/main/java/org/niis/xroad/common/managementrequest/model/AuthCertRegRequest.java
@@ -4,17 +4,17 @@
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
  * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,8 +23,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package ee.ria.xroad.common.request;
+package org.niis.xroad.common.managementrequest.model;
 
+import ee.ria.xroad.common.CodedException;
 import ee.ria.xroad.common.SystemProperties;
 import ee.ria.xroad.common.identifier.ClientId;
 import ee.ria.xroad.common.message.SoapMessageImpl;
@@ -33,6 +34,8 @@ import ee.ria.xroad.common.util.MimeTypes;
 import ee.ria.xroad.signer.protocol.SignerClient;
 import ee.ria.xroad.signer.protocol.dto.CertificateInfo;
 import ee.ria.xroad.signer.protocol.dto.MemberSigningInfo;
+import ee.ria.xroad.signer.protocol.message.GetKeyIdForCertHash;
+import ee.ria.xroad.signer.protocol.message.GetKeyIdForCertHashResponse;
 import ee.ria.xroad.signer.protocol.message.GetMemberSigningInfo;
 import ee.ria.xroad.signer.protocol.message.Sign;
 import ee.ria.xroad.signer.protocol.message.SignResponse;
@@ -46,28 +49,32 @@ import java.io.IOException;
 import java.io.InputStream;
 
 import static ee.ria.xroad.common.ErrorCodes.X_CANNOT_CREATE_SIGNATURE;
+import static ee.ria.xroad.common.ErrorCodes.X_CERT_VALIDATION;
 import static ee.ria.xroad.common.ErrorCodes.translateException;
 import static ee.ria.xroad.common.ErrorCodes.translateWithPrefix;
 import static ee.ria.xroad.common.util.CryptoUtils.calculateDigest;
+import static ee.ria.xroad.common.util.CryptoUtils.readCertificate;
 import static ee.ria.xroad.common.util.MimeUtils.HEADER_SIG_ALGO_ID;
 import static ee.ria.xroad.common.util.MimeUtils.mpRelatedContentType;
 
 @Slf4j
-abstract class GenericClientRequest implements ManagementRequest {
+public class AuthCertRegRequest implements ManagementRequest {
     private static final String SIGNATURE_DIGEST_ALGORITHM_ID =
             SystemProperties.getAuthCertRegSignatureDigestAlgorithmId();
 
-    private final ClientId client;
+    private final byte[] authCert;
+    private final ClientId owner;
     private final SoapMessageImpl requestMessage;
 
-    private CertificateInfo clientCert;
+    private CertificateInfo ownerCert;
 
     private byte[] dataToSign;
 
     private MultiPartOutputStream multipart;
 
-    GenericClientRequest(ClientId client, SoapMessageImpl request) throws Exception {
-        this.client = client;
+    public AuthCertRegRequest(byte[] authCert, ClientId owner, SoapMessageImpl request) throws Exception {
+        this.authCert = authCert;
+        this.owner = owner;
         this.requestMessage = request;
 
         this.dataToSign = request.getBytes();
@@ -90,37 +97,58 @@ abstract class GenericClientRequest implements ManagementRequest {
 
     @Override
     public InputStream getRequestContent() throws Exception {
+        verifyAuthCert();
+
         ByteArrayOutputStream out = new ByteArrayOutputStream();
         multipart = new MultiPartOutputStream(out);
 
         writeSoap();
-        writeSignature();
-        writeCert();
+        writeSignatures();
+        writeCerts();
 
         multipart.close();
         return new ByteArrayInputStream(out.toByteArray());
     }
 
-    private void writeCert() throws Exception {
-        // Write client certificate and corresponding OCSP response
-        multipart.startPart(MimeTypes.BINARY);
-        multipart.write(clientCert.getCertificateBytes());
-        multipart.startPart(MimeTypes.BINARY);
-        multipart.write(clientCert.getOcspBytes());
+    private void verifyAuthCert() throws Exception {
+        try {
+            readCertificate(authCert).checkValidity();
+        } catch (Exception e) {
+            throw new CodedException(X_CERT_VALIDATION, "Authentication certificate is invalid: %s", e.getMessage());
+        }
     }
 
-    private void writeSignature() throws Exception {
+    private void writeCerts() throws Exception {
+        // Write authentication certificate
+        multipart.startPart(MimeTypes.BINARY);
+        multipart.write(authCert);
+
+        // Write security server owner certificate and corresponding OCSP response
+        multipart.startPart(MimeTypes.BINARY);
+        multipart.write(ownerCert.getCertificateBytes());
+        multipart.startPart(MimeTypes.BINARY);
+        multipart.write(ownerCert.getOcspBytes());
+    }
+
+    private void writeSignatures() throws Exception {
+        GetKeyIdForCertHashResponse authKeyId = getAuthKeyId();
         MemberSigningInfo memberSigningInfo = getMemberSigningInfo();
 
-        String clientSignAlgoId = CryptoUtils.getSignatureAlgorithmId(SIGNATURE_DIGEST_ALGORITHM_ID,
+        String authKeySignAlogId = CryptoUtils.getSignatureAlgorithmId(SIGNATURE_DIGEST_ALGORITHM_ID,
+                authKeyId.getSignMechanismName());
+        String ownerSignAlgoId = CryptoUtils.getSignatureAlgorithmId(SIGNATURE_DIGEST_ALGORITHM_ID,
                 memberSigningInfo.getSignMechanismName());
 
-        String[] clientSignaturePartHeaders = {HEADER_SIG_ALGO_ID + ": " + clientSignAlgoId};
+        String[] authSignaturePartHeaders = {HEADER_SIG_ALGO_ID + ": " + authKeySignAlogId};
+        String[] ownerSignaturePartHeaders = {HEADER_SIG_ALGO_ID + ": " + ownerSignAlgoId};
 
         byte[] digest = calculateDigest(SIGNATURE_DIGEST_ALGORITHM_ID, dataToSign);
 
-        multipart.startPart(MimeTypes.BINARY, clientSignaturePartHeaders);
-        multipart.write(createSignature(memberSigningInfo.getKeyId(), clientSignAlgoId, digest));
+        multipart.startPart(MimeTypes.BINARY, authSignaturePartHeaders);
+        multipart.write(createSignature(authKeyId.getKeyId(), authKeySignAlogId, digest));
+
+        multipart.startPart(MimeTypes.BINARY, ownerSignaturePartHeaders);
+        multipart.write(createSignature(memberSigningInfo.getKeyId(), ownerSignAlgoId, digest));
     }
 
     private void writeSoap() throws IOException {
@@ -128,11 +156,21 @@ abstract class GenericClientRequest implements ManagementRequest {
         multipart.write(dataToSign);
     }
 
+    private GetKeyIdForCertHashResponse getAuthKeyId() {
+        try {
+            String certHash = CryptoUtils.calculateCertHexHash(authCert);
+
+            return SignerClient.execute(new GetKeyIdForCertHash(certHash));
+        } catch (Exception e) {
+            throw translateException(e);
+        }
+    }
+
     private MemberSigningInfo getMemberSigningInfo() throws Exception {
         try {
-            MemberSigningInfo signingInfo = SignerClient.execute(new GetMemberSigningInfo(client));
+            MemberSigningInfo signingInfo = SignerClient.execute(new GetMemberSigningInfo(owner));
 
-            clientCert = signingInfo.getCert();
+            ownerCert = signingInfo.getCert();
 
             return signingInfo;
         } catch (Exception e) {
