+++ src/common-management-request/src/main/java/org/niis/xroad/common/managementrequest/model/GenericClientRequest.java
@@ -23,9 +23,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package ee.ria.xroad.common.request;
+package org.niis.xroad.common.managementrequest.model;
 
-import ee.ria.xroad.common.CodedException;
 import ee.ria.xroad.common.SystemProperties;
 import ee.ria.xroad.common.identifier.ClientId;
 import ee.ria.xroad.common.message.SoapMessageImpl;
@@ -34,8 +33,6 @@ import ee.ria.xroad.common.util.MimeTypes;
 import ee.ria.xroad.signer.protocol.SignerClient;
 import ee.ria.xroad.signer.protocol.dto.CertificateInfo;
 import ee.ria.xroad.signer.protocol.dto.MemberSigningInfo;
-import ee.ria.xroad.signer.protocol.message.GetKeyIdForCertHash;
-import ee.ria.xroad.signer.protocol.message.GetKeyIdForCertHashResponse;
 import ee.ria.xroad.signer.protocol.message.GetMemberSigningInfo;
 import ee.ria.xroad.signer.protocol.message.Sign;
 import ee.ria.xroad.signer.protocol.message.SignResponse;
@@ -49,32 +46,28 @@ import java.io.IOException;
 import java.io.InputStream;
 
 import static ee.ria.xroad.common.ErrorCodes.X_CANNOT_CREATE_SIGNATURE;
-import static ee.ria.xroad.common.ErrorCodes.X_CERT_VALIDATION;
 import static ee.ria.xroad.common.ErrorCodes.translateException;
 import static ee.ria.xroad.common.ErrorCodes.translateWithPrefix;
 import static ee.ria.xroad.common.util.CryptoUtils.calculateDigest;
-import static ee.ria.xroad.common.util.CryptoUtils.readCertificate;
 import static ee.ria.xroad.common.util.MimeUtils.HEADER_SIG_ALGO_ID;
 import static ee.ria.xroad.common.util.MimeUtils.mpRelatedContentType;
 
 @Slf4j
-class AuthCertRegRequest implements ManagementRequest {
+abstract class GenericClientRequest implements ManagementRequest {
     private static final String SIGNATURE_DIGEST_ALGORITHM_ID =
             SystemProperties.getAuthCertRegSignatureDigestAlgorithmId();
 
-    private final byte[] authCert;
-    private final ClientId owner;
+    private final ClientId client;
     private final SoapMessageImpl requestMessage;
 
-    private CertificateInfo ownerCert;
+    private CertificateInfo clientCert;
 
     private byte[] dataToSign;
 
     private MultiPartOutputStream multipart;
 
-    AuthCertRegRequest(byte[] authCert, ClientId owner, SoapMessageImpl request) throws Exception {
-        this.authCert = authCert;
-        this.owner = owner;
+    GenericClientRequest(ClientId client, SoapMessageImpl request) throws Exception {
+        this.client = client;
         this.requestMessage = request;
 
         this.dataToSign = request.getBytes();
@@ -97,58 +90,37 @@ class AuthCertRegRequest implements ManagementRequest {
 
     @Override
     public InputStream getRequestContent() throws Exception {
-        verifyAuthCert();
-
         ByteArrayOutputStream out = new ByteArrayOutputStream();
         multipart = new MultiPartOutputStream(out);
 
         writeSoap();
-        writeSignatures();
-        writeCerts();
+        writeSignature();
+        writeCert();
 
         multipart.close();
         return new ByteArrayInputStream(out.toByteArray());
     }
 
-    private void verifyAuthCert() throws Exception {
-        try {
-            readCertificate(authCert).checkValidity();
-        } catch (Exception e) {
-            throw new CodedException(X_CERT_VALIDATION, "Authentication certificate is invalid: %s", e.getMessage());
-        }
+    private void writeCert() throws Exception {
+        // Write client certificate and corresponding OCSP response
+        multipart.startPart(MimeTypes.BINARY);
+        multipart.write(clientCert.getCertificateBytes());
+        multipart.startPart(MimeTypes.BINARY);
+        multipart.write(clientCert.getOcspBytes());
     }
 
-    private void writeCerts() throws Exception {
-        // Write authentication certificate
-        multipart.startPart(MimeTypes.BINARY);
-        multipart.write(authCert);
-
-        // Write security server owner certificate and corresponding OCSP response
-        multipart.startPart(MimeTypes.BINARY);
-        multipart.write(ownerCert.getCertificateBytes());
-        multipart.startPart(MimeTypes.BINARY);
-        multipart.write(ownerCert.getOcspBytes());
-    }
-
-    private void writeSignatures() throws Exception {
-        GetKeyIdForCertHashResponse authKeyId = getAuthKeyId();
+    private void writeSignature() throws Exception {
         MemberSigningInfo memberSigningInfo = getMemberSigningInfo();
 
-        String authKeySignAlogId = CryptoUtils.getSignatureAlgorithmId(SIGNATURE_DIGEST_ALGORITHM_ID,
-                authKeyId.getSignMechanismName());
-        String ownerSignAlgoId = CryptoUtils.getSignatureAlgorithmId(SIGNATURE_DIGEST_ALGORITHM_ID,
+        String clientSignAlgoId = CryptoUtils.getSignatureAlgorithmId(SIGNATURE_DIGEST_ALGORITHM_ID,
                 memberSigningInfo.getSignMechanismName());
 
-        String[] authSignaturePartHeaders = {HEADER_SIG_ALGO_ID + ": " + authKeySignAlogId};
-        String[] ownerSignaturePartHeaders = {HEADER_SIG_ALGO_ID + ": " + ownerSignAlgoId};
+        String[] clientSignaturePartHeaders = {HEADER_SIG_ALGO_ID + ": " + clientSignAlgoId};
 
         byte[] digest = calculateDigest(SIGNATURE_DIGEST_ALGORITHM_ID, dataToSign);
 
-        multipart.startPart(MimeTypes.BINARY, authSignaturePartHeaders);
-        multipart.write(createSignature(authKeyId.getKeyId(), authKeySignAlogId, digest));
-
-        multipart.startPart(MimeTypes.BINARY, ownerSignaturePartHeaders);
-        multipart.write(createSignature(memberSigningInfo.getKeyId(), ownerSignAlgoId, digest));
+        multipart.startPart(MimeTypes.BINARY, clientSignaturePartHeaders);
+        multipart.write(createSignature(memberSigningInfo.getKeyId(), clientSignAlgoId, digest));
     }
 
     private void writeSoap() throws IOException {
@@ -156,21 +128,11 @@ class AuthCertRegRequest implements ManagementRequest {
         multipart.write(dataToSign);
     }
 
-    private GetKeyIdForCertHashResponse getAuthKeyId() {
-        try {
-            String certHash = CryptoUtils.calculateCertHexHash(authCert);
-
-            return SignerClient.execute(new GetKeyIdForCertHash(certHash));
-        } catch (Exception e) {
-            throw translateException(e);
-        }
-    }
-
     private MemberSigningInfo getMemberSigningInfo() throws Exception {
         try {
-            MemberSigningInfo signingInfo = SignerClient.execute(new GetMemberSigningInfo(owner));
+            MemberSigningInfo signingInfo = SignerClient.execute(new GetMemberSigningInfo(client));
 
-            ownerCert = signingInfo.getCert();
+            clientCert = signingInfo.getCert();
 
             return signingInfo;
         } catch (Exception e) {
