+++ src/common-management-request/src/test/java/org/niis/xroad/common/managementrequest/ManagementRequestSenderTest.java
@@ -1,5 +1,6 @@
 /**
  * The MIT License
+ *
  * Copyright (c) 2019- Nordic Institute for Interoperability Solutions (NIIS)
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
@@ -23,36 +24,33 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package ee.ria.xroad.common.request;
+package org.niis.xroad.common.managementrequest;
 
 import ee.ria.xroad.common.message.SoapMessageImpl;
 
-import java.io.InputStream;
+import org.junit.Test;
+
+import static ee.ria.xroad.common.message.SoapMessageTestUtil.createResponse;
+import static junit.framework.Assert.assertEquals;
 
 /**
- * Describes the management request to be sent.
+ * Testing static methods of ManagementRequestSender
  */
-public interface ManagementRequest {
+public class ManagementRequestSenderTest {
 
     /**
-     * @return the request SOAP message
+     * Tests whether getting request id works as intended.
+     * @throws Exception in case of unexpected errors
      */
-    SoapMessageImpl getRequestMessage();
+    @Test
+    public void getRequestIdFromManagementServiceResponse() throws Exception {
+        SoapMessageImpl response =
+                createResponse("response-with-requestId.answer");
 
-    /**
-     * @return the request content
-     * @throws Exception if an error occurs
-     */
-    InputStream getRequestContent() throws Exception;
+        Integer requestId = ManagementRequestSender.getRequestId(response);
 
-    /**
-     * @return the request content type (such as text/xml)
-     */
-    String getRequestContentType();
-
-    /**
-     * @return the response content type
-     */
-    String getResponseContentType();
+        Integer expectedRequestId = 413;
+        assertEquals(expectedRequestId, requestId);
+    }
 
 }
