+++ src/proxy-ui-api/build.gradle
@@ -31,6 +31,7 @@ dependencies {
     implementation project(':serverconf')
     implementation project(':common-ui')
     implementation project(':common-rest-api')
+    implementation project(':common-management-request')
 
     implementation('org.springframework.boot:spring-boot-starter-security')
     implementation('org.springframework.boot:spring-boot-starter-web')
