+++ src/settings.gradle
@@ -15,6 +15,7 @@ include "serverconf"
 include "monitoring-conf"
 include "common-op-monitoring"
 include "common-rest-api"
+include 'common-management-request'
 
 // Main projects
 include "proxy"
